# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hangman Server API",
    "description": "\n## Hangman Server API\n\nREST API pentru jocul Hangman cu suport pentru:\n- Autentificare utilizatori (JWT)\n- Sesiuni de joc (1, 100 sau N jocuri)\n- Statistici și leaderboard\n- Administrare dicționare (admin)\n- GDPR compliance (ștergere cont, export date)\n- Rate limiting\n- Paginare cu Link headers (RFC 5988)\n\n### Autentificare\n\nMajoritatea endpoint-urilor necesită autentificare prin JWT token în header:\n```\nAuthorization: Bearer <token>\n```\n\nToken-ul se obține prin POST /api/v1/auth/login.\n\n### Rate Limiting\n\n- General: 100 req/min per token\n- Sesiuni: 10 creări/min per utilizator  \n- Jocuri: 5 creări/min per sesiune\n\n### Paginare\n\nEndpoint-urile paginate returnează Link headers conform RFC 5988:\n```\nLink: </api/v1/sessions/{id}/games?page=1>; rel=\"first\", \n      </api/v1/sessions/{id}/games?page=5>; rel=\"last\",\n      </api/v1/sessions/{id}/games?page=3>; rel=\"next\"\n```\n\n### GDPR Compliance\n\n- DELETE /api/v1/users/me - Ștergere cont (Article 17)\n- GET /api/v1/users/me/export - Export date (Article 20)\n",
    "version": "1.0.0",
    "contact": {
      "name": "Hangman API Support",
      "email": "support@hangman-api.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "summary": "List User Sessions",
        "description": "List all sessions for the current user.",
        "operationId": "list_user_sessions_api_v1_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Session",
        "description": "Create a new game session.\n\nSupports idempotency via Idempotency-Key header.",
        "operationId": "create_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateSessionRequest": {
        "properties": {
          "num_games": {
            "type": "integer",
            "title": "Num Games",
            "default": 100
          },
          "dictionary_id": {
            "type": "string",
            "title": "Dictionary Id",
            "default": "dict_ro_basic"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "normal",
              "hard",
              "auto"
            ],
            "title": "Difficulty",
            "default": "auto"
          },
          "language": {
            "type": "string",
            "enum": [
              "ro",
              "en"
            ],
            "title": "Language",
            "default": "ro"
          },
          "max_misses": {
            "type": "integer",
            "title": "Max Misses",
            "default": 6
          },
          "allow_word_guess": {
            "type": "boolean",
            "title": "Allow Word Guess",
            "default": true
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "type": "object",
        "title": "CreateSessionRequest"
      },
      "ErrorResponse": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message",
          "timestamp"
        ],
        "title": "ErrorResponse",
        "description": "Standardized error response format.",
        "example": {
          "detail": "Cannot make guesses on a completed game",
          "error_code": "GAME_3003",
          "message": "Game already finished",
          "path": "/api/v1/sessions/s_1/games/g_1/guess",
          "request_id": "req_abc123",
          "timestamp": "2025-11-02T12:34:56.789Z"
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}